requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv

pyyaml
openai
requests

# Optional: Database for caching
sqlalchemy
aiosqlite
